add new page to add log configs
