--- monitoring-dashboard/components/org.wso2.micro.integrator.dashboard.web/web-app/src/commons/AddGlobalConfigsButton.js
@@ -1,38 +1,0 @@
-/*
- * Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- *
- *
- */
-
-import React from 'react';
-import { Button } from '@material-ui/core';
-import { Link } from 'react-router-dom'
-
-export default function AddGlobalConfigsButton(props) {
-    const { pageId } = props;
-
-
-    if (pageId === 'users') {
-        return <div>
-            <Button component={Link} to="/users/add" variant="contained" color="primary">
-                Add User
-            </Button>
-                </div>
-    } else {
-        return <div />
-    } 
-}
