+++ monitoring-dashboard/components/org.wso2.micro.integrator.dashboard.web/web-app/src/commons/EnhancedTable.js
@@ -27,7 +27,6 @@ import TableContainer from '@material-ui/core/TableContainer';
 import TablePagination from '@material-ui/core/TablePagination';
 import EnhancedTableHead from './TableHeaderCreater';
 import TableRowCreator from './TableRowCreator';
-import AddGlobalConfigsButton from './AddGlobalConfigsButton'
 
 export default function EnhancedTable(props) {
     const { pageInfo, dataSet } = props;
@@ -57,7 +56,6 @@ export default function EnhancedTable(props) {
 
     return (
         <div className={classes.root}>
-            <AddGlobalConfigsButton pageId={pageInfo.pageId}/>
             <Paper className={classes.paper}>
                 <TableContainer>
                     <Table
